### 2021-04-30 #3230 @whedon
[PRE REVIEW]: libfmp: A Python Package for Fundamentals of Music Processing
**Submitting author:** @fzalkow (<a href="http://orcid.org/0000-0003-1383-4541">Frank Zalkow</a>)
**Repository:** <a href="https://github.com/meinardmueller/libfmp" target ="_blank">https://github.com/meinardmueller/libfmp</a>
**Version:** v1.1.2
