### 2021-05-06 510849b
relax version requirements of dependencies
