+++ setup.py
@@ -24,17 +24,17 @@ setup(
     ],
     keywords='audio music sound',
     license='MIT',
-    install_requires=['ipython == 7.8.*',
-                      'librosa == 0.8.*',
-                      'matplotlib == 3.1.*',
-                      'music21 == 5.7.*',
-                      'numba == 0.51.*',
-                      'numpy == 1.17.*',
-                      'pandas == 1.0.*',
-                      'pretty_midi == 0.2.*',
-                      'pysoundfile == 0.9.*',
-                      'scipy == 1.3.*'],
-    python_requires='>=3.6, <=3.8',
+    install_requires=['ipython >= 7.8.0, < 8.0.0',
+                      'librosa >= 0.8.0, < 1.0.0',
+                      'matplotlib >= 3.1.0, < 4.0.0',
+                      'music21 >= 5.7.0, < 6.0.0',
+                      'numba >= 0.51.0, < 1.0.0',
+                      'numpy >= 1.17.0, < 2.0.0',
+                      'pandas >= 1.0.0, < 2.0.0',
+                      'pretty_midi >= 0.2.0, < 1.0.0',
+                      'pysoundfile >= 0.9.0, < 1.0.0',
+                      'scipy >= 1.3.0, < 2.0.0'],
+    python_requires='>=3.6',
     extras_require={
         'tests': ['pytest == 6.2.*']
     }
