+++ libfmp/c8/c8s1_hps.py
@@ -91,7 +91,7 @@ def make_integer_odd(n):
     Returns:
         n (int): Odd integer
     """
-    if(n % 2 == 0):
+    if n % 2 == 0:
         n += 1
     return n
 
