### 2021-06-25 meinardmueller/libfmp#1 @expectopatronum
Automated tests
Hi everyone,
I am currently reviewing your JOSS submission [here](https://github.com/openjournals/joss-reviews/issues/3326).

I don't think I can tick "Automated tests" yet. I found the script "test_examples.py" and if I executed it, nothing happens. I expect that this means there are no errors. However, it would be great to have some output if everything is alright, e.g. using [unittest](https://docs.python.org/3/library/unittest.html) could help with that.
And another minor remark, it would be great to have a (sub)section in the README that instructs the user about how the tests should be executed, and in the ideal case also describes what is covered by the tests, because for me it looks like only one function per subpackage is covered.

Best regards
Verena
