### 2021-06-29 24fa1d3
hope size => hop size (see issue #2)
