### 2021-06-29 d10045f
README: section on tests (see issue #1)
