### 2021-06-29 meinardmueller/libfmp#1 @fzalkow
Dear Verena,

Many thanks for your review and the feedback. We added a section in the README that describes our philosophy regarding tests.

libfmp is special in the sense that it originates from the FMP notebooks.

https://www.audiolabs-erlangen.de/FMP

There, the correct input-output behavior of the functions is described together with extensive explanations. We consider this an "enriched" replacement for traditional unit tests. We are aware that this corresponds to a somewhat "manual" testing procedure. As far as I understand the JOSS review checklists, manual testing steps are also acceptable. We do not plan to provide comprehensive automated unit tests.

The small test script can be executed with [pytest](https://pytest.org/) (instead of [unittest](https://docs.python.org/3/library/unittest.html)) as described in the new README section. Librosa has inspired us to use this package for testing.

Best,

Frank
