### 2021-07-01 meinardmueller/libfmp#1 @expectopatronum
Hi Frank,
thanks a lot for explaining your design choices regarding testing! Sounds good to me!

> The small test script can be executed with pytest (instead of unittest) as described in the new README section. Librosa has inspired us to use this package for testing.

I wasn't aware of pytest, thanks for the additional description. Works for me!

Best regards
Verena
