### 2021-07-01 #3326 @expectopatronum
This is a more general question about how to understand "Substantial scholarly effort". If I would follow the list [here](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort), I would have to say no because:

"positive" answers:
-     **Total lines of code (LOC). Submissions under 1000 LOC will usually be flagged, those under 300 LOC will be desk rejected.:** many :)
-     **Whether the software is sufficiently useful that it is likely to be cited by your peer group.:** probably yes. It has lots of useful functions and reference implementations.

"negative" answers (leading to saying no):
-     **Age of software (is this a well-established software project) / length of commit history.:** according to the git history it's 6 month old
-     **Number of commits.:** 12
-     **Number of authors.:** 2
-     **Whether the software has already been cited in academic papers.:** My search for "libfmp" in Google Scholar only returned other work by the same group.

The two "positive" answers might outweigh some of the negatives. I don't think that the age of the software / the number of commits / number of authors should have much to say, as I would say that also one person could create a very useful software package, and 10 people might not; and additionally, the age of the git repository and the number of commits might not reflect the actual development effort as many software packages start in a private repository.

This leaves me with the last question (according to my reordered list), which is actually important. As mentioned above "libfmp" is only mentioned in papers by the same group. Since libfmp has only been available via pip for a couple of months, and was not promoted before (I think), this seems reasonable. So I also checked the [papers citing the FMP notebooks paper](https://scholar.google.com/scholar?cites=10367164936759895611&as_sdt=2005&sciodt=0,5&hl=en) which I would expect people to cite if they were using the code and it looks like the FMP notebooks paper is exclusively cited by work from the same group.

Maybe @arfon can clarify how this question should be answered given the information I provided. Thanks!
