### 2021-07-01 meinardmueller/libfmp#3 @expectopatronum
Minor issues in the documentation
Hi again,
while reviewing I noticed a few minor things in the README/documentation:
* I don't see a statement of need in the documentation. There is a detailed statement of need in the paper but the review check list also asks for a statement of need in the documentation.
* The info can be found from setup.py but for me as a user it would be helpful to see at one glance which Python version is required (e.g. in the README in section "Installing")
* Community Guidelines: There is a section regarding contributing, but it does not mention how to "Report issues or problems with the software" or "Seek support", although I would expect it through the same channels (mail to the main author, github issue).

Best regards
Verena
