### 2021-07-01 #3326 @expectopatronum
I created [an issue](https://github.com/meinardmueller/libfmp/issues) for almost all the things I noticed. There are two other things that I am not sure about and might be interesting to discuss with the other reviewer or the editor:

* The "Statement of Need" in the paper is quite long. From the sentence starting at line 58 onward it feels more like a description of the actual package rather than a statement of need, maybe this could be moved to a separate section.
* There are some design choices I am not completely happy with:
  * The paper claims to integrated and build upon the `librosa` package, but many variable are named differently, and unfortunately some are harder to understand, such as `H` instead of `hop_length`, `N` instead of `n_fft`.
  * Naming of the sub packages. I understand that the reason for naming the sub packages b, c1, ... c8 is to follow the FMP notebooks, and the book they are based on but for me as a user it is a bit cumbersome to remember or look up the sub package everytime instead of remembering a meaningful package name. I think it might be possible to keep the structure as it is and add some imports to the `__init__.py` files, to "rename" how the sub packages could be imported.

Would be interesting to hear other people's opinion about those two things.

Best regards
Verena
