### 2021-07-01 #3326 @fzalkow
Many thanks for the thorough review process. I just want to clarify two of the mentioned points.

Age of software: The development of libfmp went hand in hand with developing the [FMP notebooks](https://www.audiolabs-erlangen.de/resources/MIR/FMP/C0/C0.html), which have been developed internally before releasing libfmp on GitHub. This development process preceded the first GitHub release by a few years. Furthermore, some of the algorithms are based on previously unpublished internal MATLAB implementations, which predate FMP by several years.

Variable names: The deviations of the variable names compared to librosa is intentional. An essential aspect of the FMP notebooks and libfmp is to establish a close connection to the [FMP book](https://www.audiolabs-erlangen.de/fau/professor/mueller/bookFMP). This connection enables students to study fundamental MIR concepts in-depth by reading a textbook along with implementations of the concepts explained in the book. Thus, the variable names reflect the mathematical notation used in the book.

Best,

Frank
