### 2021-07-01 #3326 @expectopatronum
> Age of software ...
That's what I assumed! (and also doesn't have an impact on my review) The first comment was more of a question for the editor, since it is not clear to me what is expected from me for answering the "Substantial scholarly effort" question, although I am inclined to answer it with yes.

> Variable names ...
Thanks for the clarification, makes sense!

If you can take care of the minor issues I raised [here](https://github.com/meinardmueller/libfmp/issues/3) I can tick all the missing boxes and can finish my review.
