### 2021-07-01 1ad2124
README: statement of need, Python version, ways to report issue or seek support (see issue #3)
