### 2021-07-01 meinardmueller/libfmp#3 @fzalkow
Many thanks for these points. I addressed them in the commit 1ad21243e3db35e2861356308e4be63acdcbd122.

Best,

Frank
