### 2021-07-06 #3326 @arfon
> Maybe @arfon can clarify how this question should be answered given the information I provided. Thanks!

Thanks for raising these questions about scholarly effort @expectopatronum – the 'signals' we suggest are just pointers you might be able to use to guide your thinking. We also say that the work should represent [three or more months of effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) which is sound like it does?

After reading the README in a little more detail myself I did have one question for the authors – could you clarify @fzalkow if this is software designed for addressing research challenges or is a teaching tool? JOSS doesn't publish software designed for teaching (we encourage authors to submit to our sister journal [JOSE](https://jose.theoj.org) for submissions that fall into this category).
