### 2021-07-06 #3326 @fzalkow
Dear arfon,

Many thanks for your question.

Indeed libfmp covers both teaching and research aspects. This combination is due to the origin of libfmp. The starting point is Müller's [FMP book](https://www.audiolabs-erlangen.de/fau/professor/mueller/bookFMP), where fundamental music information retrieval (MIR) algorithms are presented in a didactically prepared textbook. These algorithms summarize the essence from the research of many years. Then, we developed the [FMP notebooks](https://www.audiolabs-erlangen.de/resources/MIR/FMP/C0/C0.html), which expand and complement the FMP book. Here, the aim is more on the teaching side. During the development of the FMP notebooks, many research-related algorithms have been collected in the libfmp package, which we now made available independent of the FMP notebooks. Thus, compared to the FMP notebooks, the research aspect is more dominant in libfmp. Due to this development, we consider libfmp an ideal tool to transition from studying (using the FMP notebooks and the FMP book) to research. The package is intended for research purposes and contains (similar to librosa) many implementations of core algorithms and advanced methods for MIR.

Concerning three or more months of effort: This holds undoubtedly. The development of libfmp started a few years ago. Considering the origins in the FMP book, it even can be traced back more than a decade.

Best,

Frank
