### 2021-07-07 #3326 @arfon
> The package is intended for research purposes and contains (similar to librosa) many implementations of core algorithms and advanced methods for MIR.

Got it. Thanks for the additional context and background.

@brunaw – do you think you might be able to complete your review soon?
