### 2021-07-09 #3326 @brunaw
@expectopatronum Just my two cents of the previous discussion. It's a little strange to me that the FMP notebooks paper isn't cited much because, at least from my experience, the notebooks are widely known in the ISMIR community. I know a few people who have used it in their research so they might have missed it when writing their references. This also leads to the age question, the first time I got to know the FMP notebooks was about ~3 years ago, and they were already very well developed/organized by them, so the software is certainly older than 6 months. I hope this helps with something.
