### 2021-07-13 #3326 @brunaw
@arfon I am unable to change/click in my checklist items, could you please help me fix it?
