### 2021-07-13 #3326 @brunaw
@danielskatz I think it was just that I didn't do anything about the first invitation and it didn't let you re-invite me even thought the invitation was already expired

Thank you! and apologies for the trouble
