### 2021-07-13 #3326 @brunaw
Dear @fzalkow,

Here are my comments/suggestions about the submission,

- I didn't find the Statement of Need too long, more often than not I see authors writing insufficient/generic Statements of Need so this was a good surprise. I do agree with @expectopatronum that the text can be separated into "Introduction" + "Statement of Need" though, it would make the writing flow better

- I found most functions/arguments descriptions very short. I understand that sometimes it seems like a 4-word sentence properly describes a function, but considering that this software will be mainly used by students and/or people who just started learning MIR, it would be nice to have more comprehensive descriptions. For  example, the `b.b_annotation.read_csv()` function is described as

> Reads a CSV file

And it could be something more complete like

> Reads a CSV file in table format and creates a pd.DataFrame from it, with observations in the rows and variables in the columns

Just a quick compliment, I think this was one of the best submissions I've reviewed so far. The paper was well written and I was able to check all boxes right away. The issues pointed by the first reviewer were well addressed, so I would like to congratulate the authors for the good work.
