### 2021-07-14 f9e8895
improved docstrings for functions without link to specific FMP notebook: 1. py files
