+++ libfmp/b/b_plot.py
@@ -21,7 +21,7 @@ FMP_COLORMAPS = {
 
 def plot_signal(x, Fs=1, T_coef=None, ax=None, figsize=(6, 2), xlabel='Time (seconds)', ylabel='', title='', dpi=72,
                 ylim=True, **kwargs):
-    """Plot a signal, e.g. a waveform or a novelty function
+    """Line plot visualization of a signal, e.g. a waveform or a novelty function.
 
     Args:
         x: Input signal
@@ -74,7 +74,7 @@ def plot_signal(x, Fs=1, T_coef=None, ax=None, figsize=(6, 2), xlabel='Time (sec
 def plot_matrix(X, Fs=1, Fs_F=1, T_coef=None, F_coef=None, xlabel='Time (seconds)', ylabel='Frequency (Hz)',
                 xlim=None, ylim=None, clim=None, title='', dpi=72,
                 colorbar=True, colorbar_aspect=20.0, cbar_label='', ax=None, figsize=(6, 3), **kwargs):
-    """Plot a matrix, e.g. a spectrogram or a tempogram
+    """2D raster visualization of a matrix, e.g. a spectrogram or a tempogram.
 
     Args:
         X: The matrix
@@ -153,7 +153,7 @@ def plot_matrix(X, Fs=1, Fs_F=1, T_coef=None, F_coef=None, xlabel='Time (seconds
 
 
 def plot_chromagram(*args, chroma_yticks=np.arange(12), **kwargs):
-    """Calls libfmp.b.plot_matrix and sets chroma labels
+    """Call libfmp.b.plot_matrix and sets chroma labels.
 
     See :func:`libfmp.b.b_plot.plot_matrix` for parameters and return values.
     """
@@ -169,7 +169,7 @@ def plot_chromagram(*args, chroma_yticks=np.arange(12), **kwargs):
 
 
 def compressed_gray_cmap(alpha=5, N=256, reverse=False):
-    """Creates a logarithmically or exponentially compressed grayscale colormap
+    """Create a logarithmically or exponentially compressed grayscale colormap.
 
     Args:
         alpha (float): The compression factor. If alpha > 0, it performs log compression (enhancing black colors).
@@ -210,7 +210,7 @@ class MultiplePlotsWithColorbar():
     """
 
     def __init__(self, num_plots, figsize=(8, 4), dpi=72, cbar_ratio=0.1, height_ratios=None):
-        """Creates an instance of the MultiplePlotsWithColorbar class
+        """Create an instance of the MultiplePlotsWithColorbar class
 
         Args:
             num_plots: Number of plots (also number of rows)
@@ -234,7 +234,7 @@ class MultiplePlotsWithColorbar():
             self.cbar_axes.append(plt.subplot(gs[i, 1]))
 
     def make_colorbars(self):
-        """Creates colorbars if the corresponding row needs a colorbar, or hides the axis in other cases."""
+        """Create colorbars if the corresponding row needs a colorbar, or hides the axis in other cases."""
         for i in range(self.num_plots):
             ax_img = self.axes[i].get_images()
 
@@ -247,7 +247,7 @@ class MultiplePlotsWithColorbar():
 
 
 def color_argument_to_dict(colors, labels_set, default='gray'):
-    """Creates a color dictionary
+    """Create a dictionary that maps labels to colors.
 
     Args:
         colors: Several options: 1. string of ``FMP_COLORMAPS``, 2. string of matplotlib colormap,
@@ -289,7 +289,7 @@ def color_argument_to_dict(colors, labels_set, default='gray'):
 
 
 def check_line_annotations(annot, default_label=''):
-    """Checks line annotation. If label is missing, adds an default label.
[…]
