### 2021-07-14 57905e0
improved docstrings for functions without link to specific FMP notebook: 2. docs
