### 2021-07-14 #3326 @meinardmueller
Dear @brunaw,

Thanks for your suggestions and comments. As for your first point, we followed your suggestion by adding one more headline, which better structures the paper.  As for the second point, we have the following suggestion. Most of the libfmp-functions are introduced step-by-step and explained in the FMP Notebooks; in the header of the libfmp-functions, we provided a link to the respective FMP notebook (provided as static html-version on our website). There are only a few libfmp-functions that do not have a companion in the FMP notebooks (e.g., the one you mentioned). For those functions, we would add some more details in the header as suggested by you. Doing this , we would require some weeks (also due to holidays). Does this sound ok for you?
