### 2021-07-14 #3326 @fzalkow
Dear @brunaw,

I just implemented Meinard's suggestion regarding your second point (meinardmueller/libfmp@f9e8895d2b529ee6f813afe7d3ecfa08517d5fb4, meinardmueller/libfmp@57905e09555259d5b8b9688934a6695f638d06ee). I am currently on holiday – so sorry for our somewhat asynchronous communication.

Meinard modified the markdown. So, @whedon generate pdf from branch paper.
