### 2021-07-14 #3326 @brunaw
@meinardmueller @fzalkow Yes, sounds good to me. I just saw the new changes in the paper and the function descriptions and it all looks good, thank you
