### 2021-07-14 #3326 @arfon
> @meinardmueller @fzalkow Yes, sounds good to me. I just saw the new changes in the paper and the function descriptions and it all looks good, thank you

:zap: great! I think based on the feedback of both of our reviewers here we're in good shape to move forward.

@fzalkow - Once you've made all of the changes you plan to could you make a new release of this software that includes the changes that have resulted from this review. Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
