### 2021-07-19 #3326 @fzalkow
Dear @arfon, we made a new release (latest: v1.2.1) and a Zenodo archive (DOI: [10.5281/zenodo.5113869](https://doi.org/10.5281/zenodo.5113869)). Please let us know if everything looks right. Many thanks.
