### 2021-07-20 #3326 @arfon
@fzalkow – could you please merge this PR which fixes your paper: https://github.com/meinardmueller/libfmp/pull/5
