### 2021-07-20 #3326 @fzalkow
> @fzalkow – could you please merge this PR which fixes your paper: [meinardmueller/libfmp#5](https://github.com/meinardmueller/libfmp/pull/5)

Done!
